execution_time_estimator: move aspect next to library
